v-token/backend/requirements.txt
adal==1.2.7
aio-pika==6.7.1
aiofiles==0.4.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aiohttp-jinja2==1.1.2
aiormq==3.3.1
amqp==5.0.6
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==1.6.6
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
billiard==3.6.4.0
bitarray==1.2.2
blake2b-py==0.1.4
bleach==4.1.0
cached-property==1.5.2
cachetools==5.0.0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
cytoolz==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.8
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-extensions==3.1.5
djangorestframework==3.12.4
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-bloom==1.0.4
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-tester==0.5.0b4
eth-typing==2.2.2
eth-utils==1.10.0
google-auth==2.5.0
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
ipykernel==6.7.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==0.24
jedi==0.18.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kombu==5.1.0
kubernetes==8.0.0
lazy-object-proxy==1.7.1
lru-dict==1.1.7
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
motor==2.2.0
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
netaddr==0.8.0
notebook==6.4.6
oauthlib==3.1.1
packaging==21.3
pamqp==2.3.0
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.20
protobuf==3.19.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py-ecc==4.1.0
py-evm==0.4.0a4
py-geth==3.5.0
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
pyethash==0.1.27
Pygments==2.10.0
PyJWT==2.2.0
pylint==1.9.2
pymongo==3.12.3
pyparsing==3.0.6
pyrsistent==0.18.0
pysha3==1.0.2
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2021.3
PyYAML==5.1.2
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rlp==2.0.1
rsa==4.8
semantic-version==2.8.5
Send2Trash==1.8.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
toolz==0.11.1
tornado==6.1
traitlets==5.1.1
trie==2.0.0a5
typing-extensions==3.10.0.2
urllib3==1.26.7
uvloop==0.14.0
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.24.0
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.0
